src/server/requirements.txt
git+https://github.com/cdknorow/modelstore.git@main
binaryornot==0.4.4
decorator==4.3.0
Django==3.2.12
docker==5.0.3
environ==1.0
iptools==0.6.1
pandas==1.4.1
pytest==6.2.4
xgboost==0.82
protobuf==3.18
tensorflow==2.10.0
keras==2.10.0
fastdtw==0.3.4
tsaug==0.2.1
chargebee==2.7.8
celery==5.2.3
gevent==22.10.2
greenlet==2.0.2
gunicorn==20.1.0
psycopg2-binary==2.9.3
redis==4.1.4
boto3==1.26.69
botocore==1.29.69
billiard==3.6.4.0
scikit-learn==1.1.0
scipy==1.8.0
numpy==1.22.4
django-cors-headers==3.11.0
django-cprofile-middleware==1.0.5
django-crispy-forms==1.14.0
django-crum==0.7.9
django-environ==0.8.1
django-filter==21.1
django-oauth-toolkit==1.7.0
django-oauth2==1.2
django-postgres-copy==2.7.0
django-querysetsequence==0.15
django-ratelimit==3.0.1
django-recaptcha==3.0.0
django-registration-redux==2.9
django-rest-framework-docs==0.1.7
django-ses==2.5.0
django-widget-tweaks==1.4.12
djangorestframework==3.13.1
djangorestframework-api-key==2.1.0
djangorestframework-csv==2.1.1
djangorestframework-jwt==1.11.0
flower==2.0.1
silabs-mltk==0.12.0
requests==2.31.0
appdirs
librosa
pyaml
wurlitzer
pytest-html
pytest-cov
pytest-xdist
pytest-django
tflite-support #mltk-required
prettytable #mltk-required
patool #mltk-required
netron #mltk-required
drf-spectacular==0.24.2
umap-learn
librosa==0.9.1
tensorflow-model-optimization==0.7.3
